fcm-service/requirements.txt
boto3
firebase-admin==7.1.0
# aws-xray-sdk

opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-botocore
opentelemetry-instrumentation-boto3
opentelemetry-propagator-aws-xray